Sequester RHIME parameter handling
