+++ openghg_inversions/rhime.py
@@ -33,7 +33,6 @@ import inspect
 from pathlib import Path
 from typing import Any, Literal, cast
 import time
-import warnings
 
 import arviz as az
 import numpy as np
@@ -44,8 +43,12 @@ from openghg.util import split_function_inputs  # pyright: ignore[reportPrivateI
 
 from openghg_inversions.array_ops import sparse_xr_dot
 from openghg_inversions.basis.basis_functions import BasisFunctions
-from openghg_inversions.config import config
-from openghg_inversions.inversion_data import RhimePreparedInputs, prepare_rhime_inputs
+from openghg_inversions import _rhime_params as rhime_params
+from openghg_inversions._rhime_params import params_from_config, resolve_flux_sources
+from openghg_inversions.inversion_data import (
+    RhimePreparedInputs as _RhimePreparedInputs,
+    prepare_rhime_inputs as _prepare_rhime_inputs,
+)
 from openghg_inversions.models import (
     DEFAULT_X_PRIOR,
     build_rhime_model,
@@ -56,7 +59,19 @@ from openghg_inversions.postprocessing.inversion_output import InversionOutput
 from openghg_inversions.utils import ncdf_encoding
 
 OutputFormat = Literal["none", "inv_out", "basic", "paris"]
-MinErrorConfig = Literal["percentile", "residual"] | dict[str, float] | None | int | float
+
+__all__ = [
+    "SectorSpec",
+    "RhimeModelSpec",
+    "RhimeOutputSpec",
+    "RhimeSamplingSpec",
+    "RhimeRunSpec",
+    "RhimeResult",
+    "params_from_config",
+    "resolve_flux_sources",
+    "run_rhime",
+    "run_rhime_multisector",
+]
 
 
 @dataclass(frozen=True)
@@ -228,264 +243,6 @@ class RhimeResult:
     model: pm.Model | None = None
     inv_out: InversionOutput | None = None
     sampling_spec: RhimeSamplingSpec = field(default_factory=RhimeSamplingSpec)
-
-
-def _as_list(value: str | Sequence[str] | None) -> list[str] | None:
-    """Convert a scalar/list-like value to a list of strings."""
-    if value is None:
-        return None
-    if isinstance(value, str):
-        return [value]
-    return [str(item) for item in value]
-
-
-def resolve_flux_sources(
-    *,
-    flux_sources: str | Sequence[str] | None = None,
-    emissions_name: str | Sequence[str] | None = None,
-) -> list[str]:
-    """Resolve new ``flux_sources`` and legacy ``emissions_name`` arguments.
-
-    Args:
-        flux_sources: Preferred RHIME field containing OpenGHG flux
-            ``source`` metadata values.
-        emissions_name: Legacy compatibility spelling accepted only when
-            ``flux_sources`` is absent.
-
-    Returns:
-        Resolved flux source names.
-
-    Raises:
-        ValueError: If no usable flux source is supplied.
-    """
-    resolved = _as_list(flux_sources)
-    if resolved is None:
-        resolved = _as_list(emissions_name)
-    if not resolved or any(source in {"", "None", "none"} for source in resolved):
-        raise ValueError("At least one flux source must be supplied via `flux_sources`.")
-    return resolved
-
-
-def params_from_config(
-    config_file: str | Path,
-    *,
-    start_date: str | None = None,
-    end_date: str | None = None,
-    output_path: str | None = None,
-    extra_kwargs: Mapping[str, Any] | None = None,
-) -> dict[str, Any]:
-    """Load RHIME run parameters from an INI config file.
-
-    Args:
-        config_file: Path to an INI configuration file.
-        start_date: Optional command-line start-date override.
-        end_date: Optional command-line end-date override.
-        output_path: Optional command-line output-path override.
-        extra_kwargs: Optional keyword overrides, normally parsed from CLI JSON.
-
-    Returns:
-        Normalized RHIME run parameters using snake-case public names.
-
-    Raises:
-        ValueError: If deprecated unsupported parameters are present or a
-            structured RHIME option has an invalid type.
-    """
-    params = dict(config.all_param(str(config_file), exclude_not_found=True, allow_new=True))
-    if start_date is not None:
-        params["start_date"] = start_date
-    if end_date is not None:
-        params["end_date"] = end_date
-    if output_path is not None:
-        params["output_path"] = output_path
-    if extra_kwargs:
-        params.update(extra_kwargs)
-    normalized = _normalise_params(params)
-    _validate_rhime_param_types(normalized)
-    return normalized
-
-
-def _normalise_params(params: Mapping[str, Any]) -> dict[str, Any]:
-    """Normalize legacy config spellings to modern snake-case names."""
-    normalized = dict(params)
-    aliases = {
-        "outputpath": "output_path",
-        "outputname": "output_name",
-        "xprior": "x_prior",
-        "bcprior": "bc_prior",
-        "sigprior": "sigma_prior",
-        "offsetprior": "offset_prior",
-        "emissions_name": "flux_sources",
-    }
-    for old, new in aliases.items():
-        if old not in normalized:
-            continue
-        if new in normalized:
-            warnings.warn(
-                f"Ignoring deprecated RHIME parameter {old!r} because {new!r} was also supplied.",
-                UserWarning,
-                stacklevel=2,
-            )
-        else:
-            warnings.warn(
-                f"RHIME parameter {old!r} is deprecated; use {new!r} instead.",
-                UserWarning,
-                stacklevel=2,
-            )
-            normalized[new] = normalized[old]
-        del normalized[old]
-
-    if "calculate_min_error" in normalized:
-        raise ValueError("`calculate_min_error` is not supported by RHIME runners; use `min_error`.")
-    if "reparameterise_log_normal" in normalized:
-        raise ValueError(
-            "`reparameterise_log_normal` is not supported by RHIME runners; "
-            "set `reparameterise` in the relevant prior dictionary if needed."
-        )
-    if "mcmc_type" in normalized:
-        raise ValueError("`mcmc_type` is not supported by RHIME runners; use `nuts_sampler` if needed.")
-
-    return normalized
-
-
-def _invalid_config_type_message(name: str, expected: str, value: Any) -> str:
-    """Build an actionable RHIME config type error."""
-    return (
-        f"Invalid RHIME config value for `{name}`: expected {expected}, "
-        f"but got {type(value).__name__}. Check braces/quotes in the .ini file."
-    )
-
-
-def _validate_mapping_option(params: Mapping[str, Any], name: str) -> None:
-    """Raise if a RHIME option is present but is not a mapping or None."""
-    if name not in params or params[name] is None:
-        return
-    if not isinstance(params[name], Mapping):
-        raise ValueError(_invalid_config_type_message(name, "a mapping/dict", params[name]))
-
-
-def _validate_rhime_param_types(params: Mapping[str, Any]) -> None:
-    """Validate structured RHIME parameter types before preparation begins."""
-    for prior_name in ("x_prior", "bc_prior", "sigma_prior", "offset_prior"):
-        _validate_mapping_option(params, prior_name)
-
-    _validate_mapping_option(params, "sector_sources")
-
-    if "sector_priors" in params and params["sector_priors"] is not None:
-        sector_priors = params["sector_priors"]
-        if not isinstance(sector_priors, Mapping):
-            raise ValueError(_invalid_config_type_message("sector_priors", "a mapping/dict", sector_priors))
-        for sector, prior in sector_priors.items():
-            if not isinstance(prior, Mapping):
-                raise ValueError(
-                    _invalid_config_type_message(
-                        f"sector_priors[{sector!r}]",
-                        "a mapping/dict",
-                        prior,
-                    )
-                )
-
-    for mapping_name in ("sampler_kwargs", "paris_postprocessing_kwargs", "offset_args", "min_error_options"):
-        _validate_mapping_option(params, mapping_name)
-
-    if "power" in params and params["power"] is not None:
-        power = params["power"]
-        if not isinstance(power, Mapping | int | float):
-            raise ValueError(_invalid_config_type_message("power", "a mapping/dict or number", power))
-
-
-def _normalise_optional_mapping(value: Mapping[str, Any] | None) -> dict[str, Any] | None:
-    """Copy an optional mapping so specs do not retain caller-owned dicts."""
-    return None if value is None else dict(value)
-
-
-def _normalise_sector_priors(
-    sector_priors: Mapping[str, Mapping[str, Any]] | None,
-) -> dict[str, dict[str, Any]] | None:
-    """Copy optional sector-prior mappings with string sector keys."""
-    if sector_priors is None:
-        return None
-    return {str(sector): dict(prior) for sector, prior in sector_priors.items()}
-
-
-def _normalise_sector_sources(
-    sector_sources: Mapping[str, Any] | None,
-) -> dict[str, str] | None:
-    """Copy optional sector-to-source mappings with string names."""
-    if sector_sources is None:
-        return None
-    return {str(sector): str(source) for sector, source in sector_sources.items()}
-
-
-def _required_run_params() -> set[str]:
-    """Return RHIME parameters required before data preparation."""
-    return {
-        "species",
-        "sites",
-        "averaging_period",
-        "domain",
-        "start_date",
-        "end_date",
-        "output_name",
-    }
-
-
-def _is_missing_required_value(value: Any) -> bool:
-    """Return true when a required RHIME parameter has no usable value."""
-    if value is None:
-        return True
-    if isinstance(value, str):
-        return not value.strip()
-    if isinstance(value, Sequence) and not isinstance(value, str | bytes) and len(value) == 0:
-        return True
-    return False
-
-
-def _validate_required_params(params: Mapping[str, Any]) -> None:
-    """Raise if normalized run parameters are missing required values."""
-    missing = [
-        name
-        for name in sorted(_required_run_params())
-        if name not in params or _is_missing_required_value(params[name])
-    ]
-    if missing:
-        raise ValueError(f"Required RHIME parameter(s) missing: {missing!r}")
-
-
-def _validate_supported_params(params: Mapping[str, Any]) -> None:
-    """Raise if normalized run parameters contain unsupported keys."""
-    data_params = set(inspect.signature(prepare_rhime_inputs).parameters)
-    runner_params = {
-        "x_prior",
-        "bc_prior",
-        "sigma_prior",
-        "offset_prior",
-        "sector_priors",
-        "sector_sources",
-        "pollution_events_from_obs",
-        "no_model_error",
-        "power",
-        "nit",
-        "burn",
-        "tune",
-        "nchain",
-        "nuts_sampler",
-        "verbose",
-        "sampler_kwargs",
-        "output_format",
-        "output_path",
-        "save_trace",
-        "save_inversion_output",
-        "paris_postprocessing_kwargs",
-        "offset_args",
-        "country_file",
-        "add_offset",
-        "sigma_per_site",
-    }
-    required = _required_run_params()
-    supported = data_params | runner_params | required
-    unsupported = sorted(set(params) - supported)
-    if unsupported:
-        raise ValueError(f"Unsupported RHIME parameter(s): {unsupported!r}")
 
 
 def _validate_output_format(output_format: str) -> None:
@@ -746,9 +503,13 @@ def _basis_matrix_for_output(
 
 
 def _materialise_basis_and_flux_for_output(
-    prepared: RhimePreparedInputs,
+    prepared: _RhimePreparedInputs,
 ) -> tuple[xr.DataArray, xr.DataArray]:
-    """Return materialised basis/flux arrays for current output adapters."""
+    """Return materialised arrays for transitional output adapters.
+
+    TODO(#383/#429): postprocessing should consume ``BasisFunctions`` directly
+    rather than requiring the runner to materialise a flat basis and flux.
+    """
     basis = _basis_matrix_for_output(
         prepared.basis_functions,
         state_dim="region",
@@ -759,7 +520,7 @@ def _materialise_basis_and_flux_for_output(
 
 def _make_inversion_output(
     *,
-    prepared: RhimePreparedInputs,
+    prepared: _RhimePreparedInputs,
     idata: az.InferenceData,
     start_date: str,
     end_date: str,
@@ -825,7 +586,7 @@ def _make_inversion_output(
 def _write_standard_outputs(
     *,
     result: RhimeResult,
-    prepared: RhimePreparedInputs,
+    prepared: _RhimePreparedInputs,
     country_file: str | None,
 ) -> None:
     """Create and optionally save standard RHIME outputs."""
@@ -911,13 +672,17 @@ def _write_standard_outputs(
             result.output_metadata["paris_flux_path"] = str(flux_file)
 
 
-def make_multisector_flux_diagnostics(
+def _make_multisector_flux_diagnostics(
     *,
     idata: az.InferenceData,
-    prepared: RhimePreparedInputs,
+    prepared: _RhimePreparedInputs,
     model_spec: RhimeModelSpec,
 ) -> xr.Dataset:
-    """Create sector-aware posterior flux diagnostics for shared-basis RHIME.
+    """Create provisional sector-aware posterior flux diagnostics.
+
+    TODO(#398/#429): replace this runner-local special case with
+    sector-aware postprocessing once the modern output layer supports multiple
+    sectors and ``BasisFunctions`` reconstruction directly.
 
     Args:
         idata: InferenceData returned by RHIME sampling.
@@ -949,10 +714,10 @@ def make_multisector_flux_diagnostics(
 def _write_multisector_outputs(
     *,
     result: RhimeResult,
-    prepared: RhimePreparedInputs,
+    prepared: _RhimePreparedInputs,
 ) -> None:
-    """Create and optionally save shared-basis multi-sector RHIME outputs."""
-    diagnostics = make_multisector_flux_diagnostics(
+    """Create and optionally save transitional multi-sector RHIME outputs."""
+    diagnostics = _make_multisector_flux_diagnostics(
         idata=result.idata,
         prepared=prepared,
         model_spec=result.model_spec,
@@ -1017,42 +782,22 @@ def _make_output_spec(
     )
 
 
-def _make_sampling_spec(
-    *,
-    nit: Any,
-    burn: Any,
-    tune: Any,
-    nchain: Any,
-    nuts_sampler: str,
-    verbose: bool,
-    sampler_kwargs: dict[str, Any] | None,
-) -> RhimeSamplingSpec:
-    """Create sampling settings from normalized RHIME parameters."""
-    return RhimeSamplingSpec(
-        nit=int(nit),
-        burn=int(burn),
-        tune=int(tune),
-        nchain=int(nchain),
-        nuts_sampler=nuts_sampler,
-        verbose=verbose,
-        sampler_kwargs=sampler_kwargs,
-    )
-
-
 def _make_rhime_runner_setup(
     *,
     params: Mapping[str, Any],
     multisector: bool,
 ) -> _RhimeRunnerSetup:
     """Normalize raw RHIME parameters into specs and preparation arguments."""
-    normalized = _normalise_params(params)
-    _validate_rhime_param_types(normalized)
-    _validate_required_params(normalized)
-    _validate_supported_params(normalized)
+    normalized = rhime_params.normalise_rhime_params(params)
+    rhime_params.validate_required_params(normalized)
+    rhime_params.validate_supported_params(
+        normalized,
+        data_params=set(inspect.signature(_prepare_rhime_inputs).parameters),
+    )
 
     remaining = dict(normalized)
     flux_sources = resolve_flux_sources(flux_sources=remaining.pop("flux_sources", None))
-    sector_sources = _normalise_sector_sources(remaining.pop("sector_sources", None))
+    sector_sources = rhime_params.normalise_sector_sources(remaining.pop("sector_sources", None))
     if not multisector and sector_sources is not None:
         raise ValueError("`sector_sources` is only supported by `run_rhime_multisector`.")
     data_flux_sources = (
@@ -1069,7 +814,7 @@ def _make_rhime_runner_setup(
         raise ValueError("`run_rhime` requires exactly one flux source.")
 
     species = remaining.pop("species")
-    sites = _as_list(remaining.pop("sites")) or []
+    sites = rhime_params.as_list(remaining.pop("sites")) or []
     domain = remaining.pop("domain")
     averaging_period = remaining.pop("averaging_period")
     start_date = remaining.pop("start_date")
@@ -1077,12 +822,12 @@ def _make_rhime_runner_setup(
     output_path = remaining.pop("output_path", None)
     output_name = remaining.pop("output_name")
 
-    x_prior = _normalise_optional_mapping(remaining.pop("x_prior", None))
-    bc_prior = _normalise_optional_mapping(remaining.pop("bc_prior", None))
-    sigma_prior = _normalise_optional_mapping(remaining.pop("sigma_prior", None))
-    offset_prior = _normalise_optional_mapping(remaining.pop("offset_prior", None))
-    sector_priors = _normalise_sector_priors(remaining.pop("sector_priors", None))
-    offset_args = _normalise_optional_mapping(remaining.get("offset_args"))
+    x_prior = rhime_params.normalise_optional_mapping(remaining.pop("x_prior", None))
+    bc_prior = rhime_params.normalise_optional_mapping(remaining.pop("bc_prior", None))
+    sigma_prior = rhime_params.normalise_optional_mapping(remaining.pop("sigma_prior", None))
+    offset_prior = rhime_params.normalise_optional_mapping(remaining.pop("offset_prior", None))
+    sector_priors = rhime_params.normalise_sector_priors(remaining.pop("sector_priors", None))
+    offset_args = rhime_params.normalise_optional_mapping(remaining.get("offset_args"))
 
     use_bc = remaining.get("use_bc", True)
     sigma_per_site = remaining.get("sigma_per_site", True)
@@ -1091,14 +836,14 @@ def _make_rhime_runner_setup(
     no_model_error = remaining.pop("no_model_error", False)
     power = remaining.pop("power", 1.99)
 
-    sampling_spec = _make_sampling_spec(
+    sampling_spec = RhimeSamplingSpec(
         nit=remaining.pop("nit", 1000),
         burn=remaining.pop("burn", 0),
         tune=remaining.pop("tune", 1000),
         nchain=remaining.pop("nchain", 4),
         nuts_sampler=remaining.pop("nuts_sampler", "pymc"),
         verbose=remaining.pop("verbose", False),
-        sampler_kwargs=_normalise_optional_mapping(remaining.pop("sampler_kwargs", None)),
+        sampler_kwargs=rhime_params.normalise_optional_mapping(remaining.pop("sampler_kwargs", None)),
     )
     output_spec = _make_output_spec(
         output_format=remaining.pop("output_format", "inv_out"),
@@ -1107,7 +852,7 @@ def _make_rhime_runner_setup(
         save_trace=remaining.pop("save_trace", False),
         save_inversion_output=remaining.pop("save_inversion_output", True),
         country_file=remaining.get("country_file"),
-        paris_postprocessing_kwargs=_normalise_optional_mapping(
+        paris_postprocessing_kwargs=rhime_params.normalise_optional_mapping(
             remaining.pop("paris_postprocessing_kwargs", None)
         ),
         multisector=multisector,
@@ -1154,14 +899,14 @@ def _make_rhime_runner_setup(
             "flux_sources": data_flux_sources,
             "split_by_sectors": multisector,
         },
-        prepare_rhime_inputs,
+        _prepare_rhime_inputs,
     )
     return _RhimeRunnerSetup(run_spec=run_spec, data_args=data_args)
 
 
 def _run_spec_with_prepared_inputs(
     run_spec: RhimeRunSpec,
-    prepared: RhimePreparedInputs,
+    prepared: _RhimePreparedInputs,
 ) -> RhimeRunSpec:
     """Update run metadata with retained sites from prepared RHIME inputs."""
     return replace(
@@ -1222,7 +967,7 @@ def _run_common(
 ) -> RhimeResult:
     """Run the shared RHIME pipeline after public wrapper/config normalization."""
     setup = _make_rhime_runner_setup(params=params, multisector=multisector)
-    prepared = prepare_rhime_inputs(**setup.data_args)
+    prepared = _prepare_rhime_inputs(**setup.data_args)
     run_spec = _run_spec_with_prepared_inputs(setup.run_spec, prepared)
 
     start_build = time.time()
